fix: funcname offset cannot start with 0
