+++ loader/funcdata_compat.go
@@ -20,7 +20,6 @@
 package loader
 
 import (
-   `encoding`
    `os`
    `unsafe`
    `sort`
@@ -172,98 +171,6 @@ type compilationUnit struct {
     fileNames []string
 }
 
-// func name table format: 
-//   nameOff[0] -> namePartA namePartB namePartC \x00 
-//   nameOff[1] -> namePartA namePartB namePartC \x00
-//  ...
-func makeFuncnameTab(funcs []Func) (tab []byte, offs []int32) {
-    offs = make([]int32, len(funcs))
-    offset := 0
-
-    for i, f := range funcs {
-        offs[i] = int32(offset)
-
-        a, b, c := funcNameParts(f.Name)
-        tab = append(tab, a...)
-        tab = append(tab, b...)
-        tab = append(tab, c...)
-        tab = append(tab, 0)
-        offset += len(a) + len(b) + len(c) + 1
-    }
-
-    return
-}
-
-// CU table format:
-//  cuOffsets[0] -> filetabOffset[0] filetabOffset[1] ... filetabOffset[len(CUs[0].fileNames)-1]
-//  cuOffsets[1] -> filetabOffset[len(CUs[0].fileNames)] ... filetabOffset[len(CUs[0].fileNames) + len(CUs[1].fileNames)-1]
-//  ...
-//
-// file name table format:
-//  filetabOffset[0] -> CUs[0].fileNames[0] \x00
-//  ...
-//  filetabOffset[len(CUs[0]-1)] -> CUs[0].fileNames[len(CUs[0].fileNames)-1] \x00
-//  ...
-//  filetabOffset[SUM(CUs,fileNames)-1] -> CUs[len(CU)-1].fileNames[len(CUs[len(CU)-1].fileNames)-1] \x00
-func makeFilenametab(cus []compilationUnit) (cutab []uint32, filetab []byte, cuOffsets []uint32) {
-    cuOffsets = make([]uint32, len(cus))
-    cuOffset := 0
-    fileOffset := 0
-
-    for i, cu := range cus {
-        cuOffsets[i] = uint32(cuOffset)
-
-        for _, name := range cu.fileNames {
-            cutab = append(cutab, uint32(fileOffset))
-
-            fileOffset += len(name) + 1
-            filetab = append(filetab, name...)
-            filetab = append(filetab, 0)
-        }
-
-        cuOffset += len(cu.fileNames)
-    }
-
-    return
-}
-
-func writeFuncdata(out *[]byte, funcs []Func) (fstart int, funcdataOffs [][]uint32) {
-    fstart = len(*out)
-    *out = append(*out, byte(0))
-    offs := uint32(1)
-
-    funcdataOffs = make([][]uint32, len(funcs))
-    for i, f := range funcs {
-
-        var writer = func(fd encoding.BinaryMarshaler) {
-            var ab []byte
-            var err error
-            if fd != nil {
-                ab, err = fd.MarshalBinary()
-                if err != nil {
-                    panic(err)
-                }
-                funcdataOffs[i] = append(funcdataOffs[i], offs)
-            } else {
-                ab = []byte{0}
-                funcdataOffs[i] = append(funcdataOffs[i], _INVALID_FUNCDATA_OFFSET)
-            }
-            *out = append(*out, ab...)
-            offs += uint32(len(ab))
-        }
-
-        writer(f.ArgsPointerMaps)
-        writer(f.LocalsPointerMaps)
-        writer(f.StackObjects)
-        writer(f.InlTree)
-        writer(f.OpenCodedDeferInfo)
-        writer(f.ArgInfo)
-        writer(f.ArgLiveInfo)
-        writer(f.WrapInfo)
-    }
-    return 
-}
-
 func makeFtab(funcs []_func, maxpc uintptr) (ftab []funcTab, pclntabSize int64, startLocations []uint32) {
     // Allocate space for the pc->func table. This structure consists of a pc offset
     // and an offset to the func structure. After that, we have a single pc
@@ -285,7 +192,6 @@ func makeFtab(funcs []_func, maxpc uintptr) (ftab []funcTab, pclntabSize int64, 
 
     // Final entry of table is just end pc offset.
     ftab = append(ftab, funcTab{maxpc, 0})
-
     return
 }
